Task07 -> refactoring;

+++ Task07_ObjC/Task07_ObjC/Sources/Library/IDPConstants.h
@@ -12,4 +12,4 @@
 
 IDPDeclareConstant(NSUInteger, IDPMaxArrayLength)
 IDPDeclareConstant(NSRange, IDPCashRange)
-IDPDeclareConstant(NSUInteger, IDPWorkTime)
+IDPDeclareConstant(unsigned int, IDPWorkTime)
